Move conftest.py so it works with uninstalled tests.
