+++ lib/matplotlib/sphinxext/tests/conftest.py
@@ -1,5 +1,5 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
-from matplotlib.tests.conftest import (mpl_test_settings,
-                                       pytest_configure, pytest_unconfigure)
+from matplotlib.testing.conftest import (mpl_test_settings,
+                                         pytest_configure, pytest_unconfigure)
